+++ CHANGELOG.rst
@@ -11,10 +11,12 @@ Features:
 Deprecation/Removals:
 
 - ``Query`` and ``QuerySelect`` fields are removed.
+- Passing of strings to ``required`` and ``allow_none`` is removed. Pass the ``error_messages`` argument instead.
 
 Support:
 
 - Add example of Schema inheritance in docs (:issue:`225`). Thanks :user:`martinstein` for the suggestion and :user:`juanrossi` for the PR.
+- Add "Customizing Error Messages" section to custom fields docs.
 
 2.1.1 (2015-10-07)
 ++++++++++++++++++